Break apart Map into Map and Marker, move icon to Marker file
